pyproject.toml
[tool.poetry]
name = "final"
version = "0.1.0"
description = ""
authors = ["Nick Lay <layns@mail.uc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-login = "^0.6.2"
flask-sqlalchemy = "^3.0.3"
uwsgi = "^2.0.21"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
